+++ packages/quicktype-core/src/language/Objective-C/constants.ts
@@ -1,0 +1,82 @@
+export const keywords = [
+    /*
+	"_Bool",
+	"_Complex",
+	"_Imaginary",
+	*/
+    "asm",
+    "atomic",
+    "auto",
+    "bool",
+    "break",
+    "case",
+    "char",
+    "const",
+    "continue",
+    "default",
+    "do",
+    "double",
+    "else",
+    "enum",
+    "extern",
+    "false",
+    "float",
+    "for",
+    "goto",
+    "if",
+    "inline",
+    "int",
+    "long",
+    "nil",
+    "nonatomic",
+    "register",
+    "restrict",
+    "retain",
+    "return",
+    "short",
+    "signed",
+    "sizeof",
+    "static",
+    "struct",
+    "switch",
+    "typedef",
+    "typeof",
+    "true",
+    "union",
+    "unsigned",
+    "void",
+    "volatile",
+    "while"
+] as const;
+
+export const forbiddenPropertyNames = [
+    "id",
+    "hash",
+    "description",
+    "init",
+    "copy",
+    "mutableCopy",
+    "superclass",
+    "debugDescription",
+    "new"
+] as const;
+
+export const booleanPrefixes = [
+    "is",
+    "are",
+    "were",
+    "was",
+    "will",
+    "all",
+    "some",
+    "many",
+    "has",
+    "have",
+    "had",
+    "does",
+    "do",
+    "requires",
+    "require",
+    "needs",
+    "need"
+] as const;
