+++ packages/quicktype-core/src/language/Objective-C/index.ts
@@ -1,0 +1,2 @@
+export { ObjectiveCTargetLanguage, objectiveCOptions } from "./language";
+export { ObjectiveCRenderer } from "./ObjectiveCRenderer";
